directory list UI improvements
