+++ src/components/TopBar/SearchBar.tsx
@@ -50,7 +50,7 @@ export default function SearchBar({
   }
 
   return (
-    <div>
+    <div className="absolute right-4 top-4">
       <Input
         value={searchValue}
         setValue={setSearchValue}
